+++ data/src/main/java/gr/eduinvoice/data/model/Lesson.kt
@@ -6,9 +6,11 @@ import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Index
 import androidx.room.PrimaryKey
+import kotlinx.serialization.Serializable
 import java.time.LocalDate
 import java.time.LocalTime
 
+@Serializable
 @Entity(
     tableName = "lessons",
     foreignKeys = [