+++ data/src/main/java/gr/eduinvoice/data/model/Student.kt
@@ -4,7 +4,9 @@ package gr.eduinvoice.data.model
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import kotlinx.serialization.Serializable
 
+@Serializable
 @Entity(tableName = "students")
 data class Student(
     @PrimaryKey(autoGenerate = true)