+++ data/src/main/java/gr/eduinvoice/data/model/StudentGroup.kt
@@ -2,8 +2,10 @@ package gr.eduinvoice.data.model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import kotlinx.serialization.Serializable
 import gr.eduinvoice.data.database.DatabaseConstants
 
+@Serializable
 @Entity(tableName = DatabaseConstants.GROUPS_TABLE)
 data class StudentGroup(
     @PrimaryKey(autoGenerate = true)